lab-1

+++ Lab1/src/main/java/models/User.java
@@ -1,5 +1,7 @@
 package models;
 
+import java.math.BigDecimal;
+
 /**
  * Класс, представляющий пользователя системы.
  * @param id уникальный идентификатор пользователя
@@ -7,4 +9,4 @@ package models;
  * @param name имя пользователя
  * @param countMoney баланс на счету
  */
-public record User(int id, String name, String password, double countMoney) { }
+public record User(int id, String name, String password, BigDecimal countMoney) { }
